python-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sentence-transformers
torch
pdfplumber
python-docx
python-pptx
chromadb
rank_bm25
groq
watchdog
numpy